requirements.txt
Flask==3.0.3
google-cloud-bigquery-storage == 2.24.0
fastavro == 1.9.4